+++ package.json
@@ -5,8 +5,9 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "start": "npm rebuild && npm run build && next start",
+    "lint": "next lint",
+    "rebuild": "npm rebuild"
   },
   "dependencies": {
     "@coral-xyz/anchor": "^0.29.0",
@@ -19,6 +20,7 @@
     "buffer": "^6.0.3",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "cookies-next": "^5.1.0",
     "lucide-react": "^0.477.0",
     "next": "15.2.1",
     "react": "^19.0.0",
